conect to chat room
